+++ eggshell/ocg/utils.py
